feat/Add device mockup showcase and extracted demo experience
