+++ package.json
@@ -16,6 +16,7 @@
     "lucide-react": "^0.542.0",
     "next": "15.5.2",
     "react": "19.1.0",
+    "react-device-frameset": "^1.3.4",
     "react-dom": "19.1.0",
     "react-markdown": "^10.1.0",
     "tailwind-merge": "^3.3.1",
