+++ src/app/demo/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import DemoExperience from "@/components/DemoExperience"
+
+export default function DemoPage() {
+  return <DemoExperience />
+}
